css/css-position/position-sticky-top.html

<!DOCTYPE html>
<title>position:sticky elements should respect the top constraint</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos" />
<meta name="assert" content="This test checks that position:sticky elements obey their top anchor after scrolling" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.scroller {
  position: relative;
  width: 100px;
  height: 200px;
  overflow: scroll;
}

.contents {
  height: 500px;
}

.prepadding {
  height: 100px;
}

.container {
  height: 200px;
}

.innerpadding {
  height: 50px;
}

.sticky {
  background-color: green;
  position: sticky;
  top: 50px;
}

.box {
  width: 100%;
  height: 100px;
}
</style>

<div id="scroller1" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="innerpadding"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<div id="scroller2" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="innerpadding"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<div id="scroller3" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="innerpadding"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<script>
test(() => {
  const scroller = document.getElementById('scroller1');
  const prepadding = scroller.querySelector('.prepadding');
  const innerpadding = scroller.querySelector('.innerpadding');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollTop = 50;

  const nonStickyTopY = prepadding.clientHeight + innerpadding.clientHeight;
  assert_equals(sticky.offsetTop, nonStickyTopY);
}, 'before reaching the sticking point the sticky box should not be offset');

test(() => {
  const scroller = document.getElementById('scroller2');
  const prepadding = scroller.querySelector('.prepadding');
  const innerpadding = scroller.querySelector('.innerpadding');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollTop = 125;

  // This math actually cancels to sticky.offsetTop == (scroller.scrollTop + 50),
  // but for clarity the calculations are left explicit.
  const nonStickyTopY = prepadding.clientHeight + innerpadding.clientHeight;
  const targetTopY = scroller.scrollTop + 50;
  const stickyOffset = targetTopY - nonStickyTopY;

  assert_equals(sticky.offsetTop, nonStickyTopY + stickyOffset);
}, 'after reaching the sticking point the sticky box should be offset');

test(() => {
  const scroller = document.getElementById('scroller3');
  const container = scroller.querySelector('.container');
  const prepadding = scroller.querySelector('.prepadding');
  const innerpadding = scroller.querySelector('.innerpadding');
  const sticky = container.querySelector('.sticky');

  scroller.scrollTop = 200;

  const nonStickyTopY = prepadding.clientHeight + innerpadding.clientHeight;
  const maxOffsetInContainer = container.clientHeight - innerpadding.clientHeight
      - sticky.clientHeight;

  assert_equals(sticky.offsetTop, nonStickyTopY + maxOffsetInContainer);
}, 'the sticky box should not be pushed outside its containing block');
</script>
